+++ game_patch/multi/alpine_packets.h
@@ -24,6 +24,7 @@ enum class af_packet_type : uint8_t
     af_damage_notify = 0x52,            // Alpine 1.1
     af_obj_update = 0x53,               // Alpine 1.1
     af_client_req = 0x55,               // Alpine 1.2
+    af_just_spawned_info = 0x56,        // Alpine 1.2
 };
 
 struct af_ping_location_req_packet
@@ -81,6 +82,24 @@ struct af_client_req_packet
     af_client_payload payload;
 };
 
+enum class af_just_spawned_info_type : uint8_t
+{
+    af_loadout = 0x00,
+};
+
+struct LoadoutEntry
+{
+    uint8_t weapon_index;
+    uint32_t ammo;
+};
+
+struct af_just_spawned_info_packet
+{
+    RF_GamePacketHeader header;
+    uint8_t info_type;  // af_just_spawned_info_type
+    uint8_t data[];     // type-specific payload
+};
+
 #pragma pack(pop)
 
 bool af_process_packet(const void* data, int len, const rf::NetAddr& addr, rf::Player* player);
@@ -96,6 +115,8 @@ void af_send_obj_update_packet(rf::Player* player);
 static void af_process_obj_update_packet(const void* data, size_t len, const rf::NetAddr& addr);
 void af_send_client_req_packet(const af_client_req_packet& packet);
 static void af_process_client_req_packet(const void* data, size_t len, const rf::NetAddr& addr);
+void af_send_just_spawned_loadout(rf::Player* to_player, std::vector<WeaponLoadoutEntry> loadout);
+static void af_process_just_spawned_info_packet(const void* data, size_t len, const rf::NetAddr& addr);
 
 // client requests
 void af_send_handicap_request(uint8_t amount);