+++ game_patch/multi/network.cpp
@@ -195,7 +195,8 @@ enum packet_type : uint8_t {
     af_ping_location       = 0x51,
     af_damage_notify       = 0x52,
     af_obj_update          = 0x53,
-    af_client_req          = 0x55
+    af_client_req          = 0x55,
+    af_just_spawned_info   = 0x56
 };
 
 // client -> server
@@ -272,6 +273,7 @@ std::array g_client_side_packet_whitelist{
     af_ping_location,
     af_damage_notify,
     af_obj_update,
+    af_just_spawned_info
 };
 // clang-format on